\name{plotWikipedia-package}
\alias{plotWikipedia-package}
\alias{plotWikipedia}
\docType{package}
\title{
\packageTitle{plotWikipedia}
}
\description{
\packageDescription{plotWikipedia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plotWikipedia}
\packageIndices{plotWikipedia}
}
\author{
\packageAuthor{plotWikipedia}

Maintainer: \packageMaintainer{plotWikipedia}
}
\references{
Laurens van der Maaten and Geoffrey Hinton (2008). Visualizing Data using t-SNE. Journal of Machine Learning Research Vol. 1, 1-48
}

\keyword{ package }
\seealso{
\code{\link{HyperLink}},
\code{\link{Word2Vec}},
\code{\link{graph.Wiki}},
\code{\link{tsne.Wiki}}
}
\examples{
ls("package:plotWikipedia")
}
