\name{Word2Vec}
\alias{Word2Vec}
\docType{data}
\title{
A dataframe containing Wikipedia data conveted by word2vec.
}
\description{
This data frame has 500 observations on the following 50 variables.
}
\usage{data("Word2Vec")}
\format{
  A data frame with 500 observations on the following 50 variables.
  \describe{
    \item{\code{Dim1}}{a numeric vector}
    \item{\code{Dim2}}{a numeric vector}
    \item{\code{Dim3}}{a numeric vector}
    \item{\code{Dim4}}{a numeric vector}
    \item{\code{Dim5}}{a numeric vector}
    \item{\code{Dim6}}{a numeric vector}
    \item{\code{Dim7}}{a numeric vector}
    \item{\code{Dim8}}{a numeric vector}
    \item{\code{Dim9}}{a numeric vector}
    \item{\code{Dim10}}{a numeric vector}
    \item{\code{Dim11}}{a numeric vector}
    \item{\code{Dim12}}{a numeric vector}
    \item{\code{Dim13}}{a numeric vector}
    \item{\code{Dim14}}{a numeric vector}
    \item{\code{Dim15}}{a numeric vector}
    \item{\code{Dim16}}{a numeric vector}
    \item{\code{Dim17}}{a numeric vector}
    \item{\code{Dim18}}{a numeric vector}
    \item{\code{Dim19}}{a numeric vector}
    \item{\code{Dim20}}{a numeric vector}
    \item{\code{Dim21}}{a numeric vector}
    \item{\code{Dim22}}{a numeric vector}
    \item{\code{Dim23}}{a numeric vector}
    \item{\code{Dim24}}{a numeric vector}
    \item{\code{Dim25}}{a numeric vector}
    \item{\code{Dim26}}{a numeric vector}
    \item{\code{Dim27}}{a numeric vector}
    \item{\code{Dim28}}{a numeric vector}
    \item{\code{Dim29}}{a numeric vector}
    \item{\code{Dim30}}{a numeric vector}
    \item{\code{Dim31}}{a numeric vector}
    \item{\code{Dim32}}{a numeric vector}
    \item{\code{Dim33}}{a numeric vector}
    \item{\code{Dim34}}{a numeric vector}
    \item{\code{Dim35}}{a numeric vector}
    \item{\code{Dim36}}{a numeric vector}
    \item{\code{Dim37}}{a numeric vector}
    \item{\code{Dim38}}{a numeric vector}
    \item{\code{Dim39}}{a numeric vector}
    \item{\code{Dim40}}{a numeric vector}
    \item{\code{Dim41}}{a numeric vector}
    \item{\code{Dim42}}{a numeric vector}
    \item{\code{Dim43}}{a numeric vector}
    \item{\code{Dim44}}{a numeric vector}
    \item{\code{Dim45}}{a numeric vector}
    \item{\code{Dim46}}{a numeric vector}
    \item{\code{Dim47}}{a numeric vector}
    \item{\code{Dim48}}{a numeric vector}
    \item{\code{Dim49}}{a numeric vector}
    \item{\code{Dim50}}{a numeric vector}
  }
}
\examples{
data(Word2Vec)
}
\keyword{datasets}
