\name{plotWikipedia-package}
\alias{plotWikipedia-package}
\alias{plotWikipedia}
\docType{package}
\title{
\packageTitle{plotWikipedia}
}
\description{
\packageDescription{plotWikipedia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plotWikipedia}
\packageIndices{plotWikipedia}
res.tsne <- tsne.Wiki(Word2Vec, dims=2)
plot(res.tsne$Y, pch=16, main="t-SNE of Pokemon GO-related Wikipedia pages")

}
\author{
\packageAuthor{plotWikipedia}

Maintainer: \packageMaintainer{plotWikipedia}
}
\references{
~~ Literature or other references for background information ~~

Visualizing Data using t-SNE
Laurens van der Maaten
Geoffrey Hinton
Journal of Machine Learning Research 1
1-48
2008
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
