\name{plotWikipedia-package}
\alias{plotWikipedia-package}
\alias{plotWikipedia}
\docType{package}
\title{
\packageTitle{plotWikipedia}
}
\description{
\packageDescription{plotWikipedia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plotWikipedia}
\packageIndices{plotWikipedia}
}
\author{
\packageAuthor{plotWikipedia}

Maintainer: \packageMaintainer{plotWikipedia}
}
\references{
Yoav Goldberg and Omer Levy (2014). word2vec Explained: deriving Mikolov et al.'s negative-sampling word-embedding method. arXiv 1402.3722

Laurens van der Maaten and Geoffrey Hinton (2008). Visualizing Data using t-SNE. Journal of Machine Learning Research Vol. 1, 1-48

Pokémon_Go
https://en.wikipedia.org/wiki/Pokémon_Go
}

\keyword{ package }
\seealso{
\code{\link{HyperLink}},
\code{\link{Word2Vec}},
\code{\link{graph.Wiki}},
\code{\link{tsne.Wiki}}
}
\examples{
ls("package:plotWikipedia")
}
